POC fix for async diff issues
Basic strategy is, when an event arrives, patch the previous render tree to match the current DOM state.

Also, since the client's specified eventHandlerId might have been superseded in the current tree, keep track of the chain of replacements.

For a real implementation, need to change the TS-side code to specify which element attribute it wants us to update (so .NET doesn't have to guess). Also I don't like the IsPatchableAttribute concept: want to replace this with actually inserting new frames into the tree, even though it's computationally expensive, because this would mean we can avoid embedding HTML-specific knowledge into the renderer.